requirements.txt
pymongo>=4.16.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.2.2
openai>=1.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
email-validator>=2.0.0